aws-eks/cluster/requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v28==2.2.0
pyyaml  